requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
bleach==5.0.1
bokeh==3.0.2
bs4==0.0.1
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
fitter==1.5.1
flatbuffers==22.10.26
fonttools==4.37.4
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-audit-log==0.2.4
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
h5py==2.10.0
html-to-json==2.0.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
joypy==0.2.6
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
logomaker==0.8
Markdown==3.4.1
MarkupSafe==2.1.1
mat73==0.59
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
ncls==0.0.66
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.3
numpy==1.23.3
oauthlib==3.2.2
openai==0.27.6
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patchworklib==0.5.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyppt==0.2.2
pyranges==0.0.120
pyrle==0.0.35
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.1.3
scikit-posthocs==0.7.0
scipy==1.4.1
seaborn==0.12.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sorted-nearest==0.0.38
soupsieve==2.3.2.post1
stack-data==0.5.1
statsmodels==0.13.5
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
UpSetPlot==0.8.0
uri-template==1.2.0
urllib3==1.26.12
utils==1.0.1
venn==0.1.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xmltodict==0.12.0
xmltojson==2.0.1
xyzservices==2022.9.0
yarl==1.9.2
zipp==3.9.0